Updated all about page components and made them responsive
