+++ src/pages/About/SubComponents/ExploreOurBusinessSection.jsx
@@ -10,25 +10,12 @@ const ExploreOurBusinessSection = ({ mapInfo }) => {
             Explore our business
           </div>
         </div>
-        <div className="w-full max-w-screen-xl mx-auto grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-2 gap-6 sm:gap-8 lg:gap-10 px-2 sm:px-4">
+        <div className="w-full max-w-screen-xl mx-auto grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-2 gap-6 sm:gap-8 lg:gap-10 px-2 sm:px-4 ">
           {mapInfo.map(elem => (
             <BusinessBranches key={elem.heading} displayElement={elem} />
           ))}
         </div>
       </div>
-
-      {/* <div className="flex flex-col mb-16">
-        <div className="flex w-full h-[80px] bg-black text-white items-center justify-center mb-[60px]">
-          <div className="text-slate-50 text-5xl font-bold font-['Poppins'] leading-[60px] ">
-            Explore our business
-          </div>
-        </div>
-        <div className="flex flex-row flex-wrap w-[1220px] h-[640px] mx-auto gap-[40px]">
-          {mapInfo.map(elem => (
-            <BusinessBranches key={elem.heading} displayElement={elem} />
-          ))}
-        </div>
-      </div> */}
     </>
   );
 };
